requirements.txt
alembic==1.0.8
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dateutil==2.8.0
python-editor==1.0.4
redis==3.2.1
requests==2.21.0
rq==0.13.0
simplifycommerce-sdk-python==1.6.0
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
